+++ src/shared/SharedConfig.h
@@ -47,7 +47,7 @@ class SharedConfig {
         };
         RecStatus recording = NO_RECORDING;
 
-        unsigned long int movie_framecount = 0;
+        unsigned int movie_framecount = 0;
 
         /* Log status */
         enum LogStatus {
